refactor(gui-v2): remove subpages of signup
